Make gzip decompression slightly more efficient
Instead of putting our compressed data into a buffer before decompressing
it I pass it into our z_stream directly. My only concer with this is, will
zlib always decompress everything it receives? Because if it doesn't, with
our next call to Decompress it won't have that data anymore. So far I haven't
seen any issues with this though.

Signed-off-by: Toon Schoenmakers <nighteyes1993@gmail.com>