Made a bunch of methods const and made Select and Insert return pointers instead
This refused to compile on travis for some reason, according to stack overflow
I'm not allowed to have DBResult as a return type as it is abstract. I am however
allowed to have DBResult* as a return type. Only disadvantage here is that you'll
have to clean up DBResult* yourself! Although this is not the case with the
NonBlockingPostGres implementation as it'll just return NULL.

Signed-off-by: Toon Schoenmakers <nighteyes1993@gmail.com>